requirements.txt
tensorflow
pillow
numpy
flask
matplotlib
